Terms

+++ app/Http/Controllers/HomeController.php
@@ -2,9 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use Illuminate\Http\Request, DB;
 use App\Models\User, App\Models\Advertisement;
-
+use App\Models\AdvertisementReview, App\Models\AdvertisementServices;
+use App\Models\AdvertisementServiceDuration, App\Models\AdvertisementWorkingDays;
+use App\Models\AdvertisementsImage,App\Models\Booking;
+use App\Models\LadyPremiumPicture, App\Models\LoveCount;
+use App\Models\ReviewLikeDislike, App\Models\UserVerification;
 
 class HomeController extends Controller
 {
@@ -22,7 +26,8 @@ class HomeController extends Controller
      * Show the application dashboard.
      *
      * @return \Illuminate\Contracts\Support\Renderable
-     */
+    */
+
     public function index()
     {
         $guard = get_guard();
@@ -33,11 +38,34 @@ class HomeController extends Controller
         }
     }
 
-    /*public function deleteMyAccount(Request $req,$userId)
+    // Delete User Account Permanently
+    public function deleteMyAccount(Request $req,$userId)
     {
-        $user = User::findOrFail($userId);
-        Advertisement::where('ladies_id',$user->id)->orWhere('club_id',$user->id)->delete();
-        $user->delete();
-        return back()->with('Success','Account Deleted SuccessFully');
-    }*/
+        DB::beginTransaction();
+        try {
+            $user = User::where('id',$userId)->first();
+            if($user){
+                $advertisement = Advertisement::where('ladies_id',$user->id)->orWhere('club_id',$user->id)->get();
+                foreach ($advertisement as $key => $adv) {
+                    AdvertisementReview::where('advertisement_id',$adv->id)->delete();
+                    AdvertisementServices::where('advertisement_id',$adv->id)->delete();
+                    AdvertisementServiceDuration::where('advertisement_id',$adv->id)->delete();
+                    AdvertisementWorkingDays::where('advertisement_id',$adv->id)->delete();
+                    AdvertisementsImage::where('advertisement_id',$adv->id)->delete();
+                    Booking::where('advertisement_id',$adv->id)->delete();
+                    ReviewLikeDislike::where('advertisement_id',$adv->id)->delete();
+                    LoveCount::where('advertisement_id',$adv->id)->delete();
+                }
+                LadyPremiumPicture::where('user_id',$user->id)->delete();
+                UserVerification::where('user_id',$user->id)->delete();
+                $user->delete();
+                DB::commit();
+                return response()->json(['error' => false,'message' => 'Account Deleted Success','data' => $req->all()]);
+            }
+            return response()->json(['error' => true,'message' => 'Invalid Information Detected','data' => $req->all()]);
+        } catch (Exception $e) {
+            DB::rollback();
+            return response()->json(['error' => true,'message' => 'Something went wrong please try after sometime']);
+        }
+    }
 }
